feat(timestamp): standardize timestamp format to i64 milliseconds across all traits and types
- Updated Account trait to use i64 timestamps instead of u64 with comprehensive documentation
- Added deprecated u64 wrapper methods for backward compatibility in Account and Funding traits
- Updated Funding trait with i64 timestamp format and detailed migration examples
- Modified all timestamp fields in funding rate related types to use i64 instead of u64
- Updated Binance exchange implementation to use i64 timestamps in fetch_my_trades and related methods
- Updated example files to demonstrate proper i64 timestamp usage throughout the library
- Enhanced documentation with timestamp format specifications and usage examples
- Added comprehensive timestamp format documentation to all relevant traits
- Updated Binance futures implementation to use i64 for funding rate history methods
- Fixed test cases to properly handle i64 timestamps and removed ignore annotations
- Added timestamp utility demonstrations in example files
- Updated Cargo.toml files with new dependencies and benchmark configurations

+++ ccxt-core/src/types/market.rs
@@ -450,15 +450,15 @@ pub struct LastPrice {
     /// Latest price.
     pub price: Decimal,
     /// Timestamp in milliseconds.
-    pub timestamp: u64,
+    pub timestamp: i64,
     /// Datetime string.
     pub datetime: String,
 }
 
 impl LastPrice {
     /// Creates a new last price instance.
-    pub fn new(symbol: String, price: Decimal, timestamp: u64) -> Self {
-        let datetime = chrono::DateTime::from_timestamp_millis(timestamp as i64)
+    pub fn new(symbol: String, price: Decimal, timestamp: i64) -> Self {
+        let datetime = chrono::DateTime::from_timestamp_millis(timestamp)
             .map(|dt| dt.format("%Y-%m-%dT%H:%M:%S%.3fZ").to_string())
             .unwrap_or_default();
